+++ src/lib/utils/scales/scale_continuous.ts
@@ -1,5 +1,7 @@
+import { bisectLeft } from 'd3-array';
 import { scaleLinear, scaleLog, scaleSqrt, scaleUtc } from 'd3-scale';
 import { DateTime } from 'luxon';
+import { clamp } from '../commons';
 import { ScaleContinuousType, ScaleType } from './scales';
 import { Scale } from './scales';
 
@@ -59,14 +61,11 @@ export function limitLogScaleDomain(domain: any[]) {
   }
   return domain;
 }
-export enum StepType {
-  StepBefore = 'before',
-  StepAfter = 'after',
-  Step = 'half',
-}
 
 export class ScaleContinuous implements Scale {
   readonly bandwidth: number;
+  readonly totalBarsInCluster: number;
+  readonly bandwidthPadding: number;
   readonly minInterval: number;
   readonly step: number;
   readonly type: ScaleType;
@@ -75,21 +74,41 @@ export class ScaleContinuous implements Scale {
   readonly isInverted: boolean;
   readonly tickValues: number[];
   readonly timeZone: string;
+  readonly barsPadding: number;
   private readonly d3Scale: any;
 
   constructor(
+    type: ScaleContinuousType,
     domain: any[],
     range: [number, number],
-    type: ScaleContinuousType,
-    clamp: boolean = false,
+    /**
+     * The desidered bandwidth for a linear band scale.
+     * @default 0
+     */
     bandwidth: number = 0,
-    /** the min interval computed on the XDomain, not available for yDomains */
+    /**
+     * The min interval computed on the XDomain. Not available for yDomains.
+     * @default 0
+     */
     minInterval: number = 0,
     /**
      * A time zone identifier. Can be any IANA zone supported by he host environment,
      * or a fixed-offset name of the form 'utc+3', or the strings 'local' or 'utc'.
+     * @default 'utc'
      */
     timeZone: string = 'utc',
+    /**
+     * The number of bars in the cluster. Used to correctly compute scales when
+     * using padding between bars.
+     * @default 1
+     */
+    totalBarsInCluster: number = 1,
+    /**
+     * The proportion of the range that is reserved for blank space between bands
+     * A number between 0 and 1.
+     * @default 0
+     */
+    barsPadding: number = 0,
   ) {
     this.d3Scale = SCALES[type]();
     if (type === ScaleType.Log) {
@@ -99,16 +118,18 @@ export class ScaleContinuous implements Scale {
       this.domain = domain;
       this.d3Scale.domain(domain);
     }
+    const safeBarPadding = clamp(barsPadding, 0, 1);
+    this.barsPadding = safeBarPadding;
+    this.bandwidth = bandwidth * (1 - safeBarPadding);
+    this.bandwidthPadding = bandwidth * safeBarPadding;
     this.d3Scale.range(range);
-    this.d3Scale.clamp(clamp);
-    // this.d3Scale.nice();
-    this.bandwidth = bandwidth;
     this.step = 0;
     this.type = type;
     this.range = range;
     this.minInterval = minInterval;
     this.isInverted = this.domain[0] > this.domain[1];
     this.timeZone = timeZone;
+    this.totalBarsInCluster = totalBarsInCluster;
     if (type === ScaleType.Time) {
       const startDomain = DateTime.fromMillis(this.domain[0], { zone: this.timeZone });
       const endDomain = DateTime.fromMillis(this.domain[1], { zone: this.timeZone });
@@ -124,7 +145,7 @@ export class ScaleContinuous implements Scale {
       });
     } else {
       if (this.minInterval > 0) {
-        const intervalCount = (this.domain[1] - this.domain[0]) / this.minInterval;
+        const intervalCount = Math.floor((this.domain[1] - this.domain[0]) / this.minInterval);
         this.tickValues = new Array(intervalCount + 1).fill(0).map((d, i) => {
           return this.domain[0] + i * this.minInterval;
         });
@@ -135,7 +156,7 @@ export class ScaleContinuous implements Scale {
   }
 
   scale(value: any) {
-    return this.d3Scale(value);
+    return this.d3Scale(value) + (this.bandwidthPadding / 2) * this.totalBarsInCluster;
   }
 
   ticks() {
@@ -148,10 +169,34 @@ export class ScaleContinuous implements Scale {
     }
     return invertedValue;
   }
-  invertWithStep(value: number, stepType?: StepType) {
-    const invertedValue = this.invert(value);
-    const forcedStep = this.bandwidth > 0 ? StepType.StepAfter : stepType;
-    return invertValue(this.domain[0], invertedValue, this.minInterval, forcedStep);
+  invertWithStep(value: number, data: number[]): any {
+    const invertedValue = this.invert(value - this.bandwidth / 2);
+    const leftIndex = bisectLeft(data, invertedValue);
+    if (leftIndex === 0) {
+      // is equal or less than the first value
+      const prevValue1 = data[leftIndex];
+      if (data.length === 0) {
+        return prevValue1;
+      }
+      const nextValue1 = data[leftIndex + 1];
+      const nextDiff1 = Math.abs(nextValue1 - invertedValue);
+      const prevDiff1 = Math.abs(invertedValue - prevValue1);
+      if (nextDiff1 < prevDiff1) {
+        return nextValue1;
+      }
+      return prevValue1;
+    }
+    if (leftIndex === data.length) {
+      return data[leftIndex - 1];
+    }
+    const nextValue = data[leftIndex];
+    const prevValue = data[leftIndex - 1];
+    const nextDiff = Math.abs(nextValue - invertedValue);
+    const prevDiff = Math.abs(invertedValue - prevValue);
+    if (nextDiff <= prevDiff) {
+      return nextValue;
+    }
+    return prevValue;
   }
 }
 
@@ -161,59 +206,4 @@ export function isContinuousScale(scale: Scale): scale is ScaleContinuous {
 
 export function isLogarithmicScale(scale: Scale) {
   return scale.type === ScaleType.Log;
-}
-
-function invertValue(
-  domainMin: number,
-  invertedValue: number,
-  minInterval: number,
-  stepType?: StepType,
-) {
-  if (minInterval > 0) {
-    switch (stepType) {
-      case StepType.StepAfter:
-        return linearStepAfter(invertedValue, minInterval);
-      case StepType.StepBefore:
-        return linearStepBefore(invertedValue, minInterval);
-      case StepType.Step:
-      default:
-        return linearStep(domainMin, invertedValue, minInterval);
-    }
-  }
-  return invertedValue;
-}
-
-/**
- * Return an inverted value that is valid from the exact point of the scale
- * till the end of the interval. |--------|********|
- * @param invertedValue the inverted value
- * @param minInterval the data minimum interval grether than 0
- */
-export function linearStepAfter(invertedValue: number, minInterval: number): number {
-  return Math.floor(invertedValue / minInterval) * minInterval;
-}
-
-/**
- * Return an inverted value that is valid from the half point before and half point
- * after the value. |----****|*****----|
- * till the end of the interval.
- * @param domainMin the domain's minimum value
- * @param invertedValue the inverted value
- * @param minInterval the data minimum interval grether than 0
- */
-export function linearStep(domainMin: number, invertedValue: number, minInterval: number): number {
-  const diff = (invertedValue - domainMin) / minInterval;
-  const base = diff - Math.floor(diff) > 0.5 ? 1 : 0;
-  return domainMin + Math.floor(diff) * minInterval + minInterval * base;
-}
-
-/**
- * Return an inverted value that is valid from the half point before and half point
- * after the value. |********|--------|
- * till the end of the interval.
- * @param invertedValue the inverted value
- * @param minInterval the data minimum interval grether than 0
- */
-export function linearStepBefore(invertedValue: number, minInterval: number): number {
-  return Math.ceil(invertedValue / minInterval) * minInterval;
 }