+++ src/lib/utils/scales/scales.test.ts
@@ -1,13 +1,14 @@
 import { DateTime } from 'luxon';
-import { limitLogScaleDomain } from './scale_continuous';
-import { createContinuousScale, createOrdinalScale, ScaleType } from './scales';
+import { ScaleBand } from './scale_band';
+import { limitLogScaleDomain, ScaleContinuous } from './scale_continuous';
+import { ScaleType } from './scales';
 
 describe('Scale Test', () => {
   test('Create an ordinal scale', () => {
     const data = ['a', 'b', 'c', 'd', 'a', 'b', 'c'];
     const minRange = 0;
     const maxRange = 100;
-    const ordinalScale = createOrdinalScale(data, minRange, maxRange);
+    const ordinalScale = new ScaleBand(data, [minRange, maxRange]);
     const { domain, range, bandwidth } = ordinalScale;
     expect(domain).toEqual(['a', 'b', 'c', 'd']);
     expect(range).toEqual([minRange, maxRange]);
@@ -25,7 +26,7 @@ describe('Scale Test', () => {
     const data = [0, 10];
     const minRange = 0;
     const maxRange = 100;
-    const linearScale = createContinuousScale(ScaleType.Linear, data, minRange, maxRange);
+    const linearScale = new ScaleContinuous(ScaleType.Linear, data, [minRange, maxRange]);
     const { domain, range } = linearScale;
     expect(domain).toEqual([0, 10]);
     expect(range).toEqual([minRange, maxRange]);
@@ -49,7 +50,7 @@ describe('Scale Test', () => {
     const data = [date1, date3];
     const minRange = 0;
     const maxRange = 100;
-    const timeScale = createContinuousScale(ScaleType.Time, data, minRange, maxRange);
+    const timeScale = new ScaleContinuous(ScaleType.Time, data, [minRange, maxRange]);
     const { domain, range } = timeScale;
     expect(domain).toEqual([date1, date3]);
     expect(range).toEqual([minRange, maxRange]);
@@ -64,7 +65,7 @@ describe('Scale Test', () => {
     const data = [1, 10];
     const minRange = 0;
     const maxRange = 100;
-    const logScale = createContinuousScale(ScaleType.Log, data, minRange, maxRange);
+    const logScale = new ScaleContinuous(ScaleType.Log, data, [minRange, maxRange]);
     const { domain, range } = logScale;
     expect(domain).toEqual([1, 10]);
     expect(range).toEqual([minRange, maxRange]);
@@ -77,7 +78,7 @@ describe('Scale Test', () => {
     const data = [0, 10];
     const minRange = 0;
     const maxRange = 100;
-    const logScale = createContinuousScale(ScaleType.Log, data, minRange, maxRange);
+    const logScale = new ScaleContinuous(ScaleType.Log, data, [minRange, maxRange]);
     const { domain, range } = logScale;
     expect(domain).toEqual([1, 10]);
     expect(range).toEqual([minRange, maxRange]);
@@ -90,7 +91,7 @@ describe('Scale Test', () => {
     const data = [0, 10];
     const minRange = 0;
     const maxRange = 100;
-    const sqrtScale = createContinuousScale(ScaleType.Sqrt, data, minRange, maxRange);
+    const sqrtScale = new ScaleContinuous(ScaleType.Sqrt, data, [minRange, maxRange]);
     const { domain, range } = sqrtScale;
     expect(domain).toEqual([0, 10]);
     expect(range).toEqual([minRange, maxRange]);
@@ -141,27 +142,26 @@ describe('Scale Test', () => {
   });
 
   test('compare ordinal scale and linear/band invertWithStep 3 bars', () => {
-    const dataLinear = [0, 2];
-    const dataOrdinal = [0, 1, 2];
+    const data = [0, 1, 2];
+    const domainLinear = [0, 2];
+    const domainOrdinal = [0, 1, 2];
     const minRange = 0;
     const maxRange = 120;
     const bandwidth = maxRange / 3;
-    const linearScale = createContinuousScale(
+    const linearScale = new ScaleContinuous(
       ScaleType.Linear,
-      dataLinear,
-      minRange,
-      maxRange - bandwidth, // we currently limit the range like that a band linear scale
+      domainLinear,
+      [minRange, maxRange - bandwidth], // we currently limit the range like that a band linear scale
       bandwidth,
-      false,
       1,
     );
-    const ordinalScale = createOrdinalScale(dataOrdinal, minRange, maxRange);
+    const ordinalScale = new ScaleBand(domainOrdinal, [minRange, maxRange]);
     expect(ordinalScale.invertWithStep(0)).toBe(0);
     expect(ordinalScale.invertWithStep(40)).toBe(1);
     expect(ordinalScale.invertWithStep(80)).toBe(2);
-    expect(linearScale.invertWithStep(0)).toBe(0);
-    expect(linearScale.invertWithStep(40)).toBe(1);
-    expect(linearScale.invertWithStep(80)).toBe(2);
+    expect(linearScale.invertWithStep(0, data)).toBe(0);
+    expect(linearScale.invertWithStep(40, data)).toBe(1);
+    expect(linearScale.invertWithStep(80, data)).toBe(2);
   });
   test('compare ordinal scale and linear/band 2 bars', () => {
     const dataLinear = [0, 1];
@@ -169,16 +169,14 @@ describe('Scale Test', () => {
     const minRange = 0;
     const maxRange = 100;
     const bandwidth = maxRange / 2;
-    const linearScale = createContinuousScale(
+    const linearScale = new ScaleContinuous(
       ScaleType.Linear,
       dataLinear,
-      minRange,
-      maxRange - bandwidth, // we currently limit the range like that a band linear scale
+      [minRange, maxRange - bandwidth], // we currently limit the range like that a band linear scale
       bandwidth,
-      false,
       1,
     );
-    const ordinalScale = createOrdinalScale(dataOrdinal, minRange, maxRange);
+    const ordinalScale = new ScaleBand(dataOrdinal, [minRange, maxRange]);
 
     expect(ordinalScale.scale(0)).toBe(0);
     expect(ordinalScale.scale(1)).toBe(50);
@@ -188,31 +186,10 @@ describe('Scale Test', () => {
     expect(ordinalScale.invertWithStep(0)).toBe(0);
     expect(ordinalScale.invertWithStep(50)).toBe(1);
     expect(ordinalScale.invertWithStep(100)).toBe(1);
-    expect(linearScale.invertWithStep(0)).toBe(0);
-    expect(linearScale.invertWithStep(50)).toBe(1);
-    // expect(linearScale.invertWithStep(100)).toBe(1);
+    expect(linearScale.invertWithStep(0, dataLinear)).toBe(0);
+    expect(linearScale.invertWithStep(50, dataLinear)).toBe(1);
+    expect(linearScale.invertWithStep(100, dataLinear)).toBe(1);
     expect(linearScale.bandwidth).toBe(50);
     expect(linearScale.range).toEqual([0, 50]);
-  });
-  test.skip('compare ordinal scale and linear/band 2 bars', () => {
-    // I'm skipping this because the current implementation of linearScale
-    // is a bit different than the ordinalScale when forcing bandwidth on linear scale
-    const dataLinear = [0, 1];
-    const dataOrdinal = [0, 1];
-    const minRange = 0;
-    const maxRange = 100;
-    const bandwidth = maxRange / 2;
-    const linearScale = createContinuousScale(
-      ScaleType.Linear,
-      dataLinear,
-      minRange,
-      maxRange - bandwidth, // we currently limit the range like that a band linear scale
-      bandwidth,
-      false,
-      1,
-    );
-    const ordinalScale = createOrdinalScale(dataOrdinal, minRange, maxRange);
-    expect(ordinalScale.invertWithStep(100)).toBe(1);
-    expect(linearScale.invertWithStep(100)).toBe(1);
   });
 });