+++ src/state/test/interactions.test.ts
@@ -153,11 +153,11 @@ describe('Chart state pointer interactions', () => {
   });
 
   test('can respond to tooltip types changes', () => {
-    store.xScale = new ScaleContinuous([0, 1], [0, 100], ScaleType.Linear, false, 50, 0.5);
+    store.xScale = new ScaleContinuous(ScaleType.Linear, [0, 1], [0, 100], 50, 0.5);
     store.yScales = new Map();
-    store.yScales.set(GROUP_ID, new ScaleContinuous([0, 1], [0, 100], ScaleType.Linear));
+    store.yScales.set(GROUP_ID, new ScaleContinuous(ScaleType.Linear, [0, 1], [0, 100]));
     store.geometriesIndex.set(0, [indexedGeom1Red]);
-
+    store.geometriesIndexKeys.push(0);
     store.tooltipType.set(TooltipType.None);
     store.setCursorPosition(10, 10 + 70);
     expect(store.tooltipData).toEqual([]);
@@ -165,11 +165,12 @@ describe('Chart state pointer interactions', () => {
 
     store.tooltipType.set(TooltipType.Follow);
     store.setCursorPosition(10, 10 + 70);
+    expect(store.geometriesIndexKeys.length).toBe(1);
     expect(store.isTooltipVisible.get()).toBe(true);
     expect(store.highlightedGeometries.length).toBe(1);
   });
 
-  describe('mouse over with ordinal scale', () => {
+  describe('mouse over with Ordinal scale', () => {
     mouseOverTestSuite(ScaleType.Ordinal);
   });
   describe('mouse over with Linear scale', () => {
@@ -198,6 +199,8 @@ function mouseOverTestSuite(scaleType: ScaleType) {
     store.yScales = yScales;
     store.geometriesIndex.set(0, [indexedGeom1Red]);
     store.geometriesIndex.set(1, [indexedGeom2Blue]);
+    store.geometriesIndexKeys.push(0);
+    store.geometriesIndexKeys.push(1);
     onOverListener = jest.fn((elements: GeometryValue[]): undefined => undefined);
     onOutListener = jest.fn((): undefined => undefined);
     store.setOnElementOverListener(onOverListener);