+++ stories/bar_chart.tsx
@@ -9,6 +9,7 @@ import {
   DARK_THEME,
   DataGenerator,
   getAxisId,
+  getGroupId,
   getSpecId,
   LIGHT_THEME,
   LineSeries,
@@ -202,6 +203,42 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={[{ x: 1, y: 2 }, { x: 2, y: 7 }, { x: 4, y: 3 }, { x: 9, y: 6 }]}
+        />
+      </Chart>
+    );
+  })
+  .add('with linear x axis no linear interval', () => {
+    return (
+      <Chart className={'story-chart'}>
+        <Settings xDomain={{ max: 100 }} />
+        <Axis
+          id={getAxisId('bottom')}
+          position={Position.Bottom}
+          title={'Bottom axis'}
+          showOverlappingTicks={true}
+        />
+        <Axis
+          id={getAxisId('left2')}
+          title={'Left axis'}
+          position={Position.Left}
+          tickFormat={(d) => Number(d).toFixed(2)}
+        />
+
+        <BarSeries
+          id={getSpecId('bars')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor="x"
+          yAccessors={['y']}
+          data={[
+            { x: 0, y: 2 },
+            { x: 10, y: 7 },
+            { x: 11.5, y: 9 },
+            { x: 13.5, y: 3 },
+            { x: 50, y: 6 },
+            { x: 66, y: 13 },
+            { x: 90, y: 4 },
+          ]}
         />
       </Chart>
     );
@@ -654,7 +691,7 @@ storiesOf('Bar Chart', module)
 
         <BarSeries
           id={getSpecId('bars1')}
-          xScaleType={ScaleType.Ordinal}
+          xScaleType={ScaleType.Linear}
           yScaleType={ScaleType.Linear}
           xAccessor="x"
           yAccessors={['y']}
@@ -752,7 +789,7 @@ storiesOf('Bar Chart', module)
   })
   .add('with high data volume', () => {
     const dg = new DataGenerator();
-    const data = dg.generateSimpleSeries(200);
+    const data = dg.generateSimpleSeries(15000);
     return (
       <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
@@ -1237,6 +1274,120 @@ storiesOf('Bar Chart', module)
           splitSeriesAccessors={[3]}
           stackAccessors={[0]}
           data={data}
+        />
+      </Chart>
+    );
+  })
+  .add('[test] switch ordinal/linear x axis', () => {
+    return (
+      <Chart className={'story-chart'}>
+        <Axis
+          id={getAxisId('bottom')}
+          position={Position.Bottom}
+          title={'Bottom axis'}
+          showOverlappingTicks={true}
+        />
+        <Axis
+          id={getAxisId('left2')}
+          title={'Left axis'}
+          position={Position.Left}
+          tickFormat={(d) => Number(d).toFixed(2)}
+        />
+
+        <BarSeries
+          id={getSpecId('bars')}
+          xScaleType={select(
+            'scaleType',
+            {
+              linear: ScaleType.Linear,
+              ordinal: ScaleType.Ordinal,
+            },
+            ScaleType.Linear,
+          )}
+          yScaleType={ScaleType.Linear}
+          xAccessor="x"
+          yAccessors={['y']}
+          data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
+        />
+      </Chart>
+    );
+  })
+  .add('stacked only grouped areas', () => {
+    const data1 = [[1, 2], [2, 2], [3, 3], [4, 5], [5, 5], [6, 3], [7, 8], [8, 2], [9, 1]];
+    const data2 = [[1, 1], [2, 2], [3, 3], [4, 4], [5, 5], [6, 4], [7, 3], [8, 2], [9, 4]];
+    const data3 = [[1, 6], [2, 6], [3, 3], [4, 2], [5, 1], [6, 1], [7, 5], [8, 6], [9, 7]];
+    const data4 = [[1, 2], [2, 6], [3, 2], [4, 9], [5, 2], [6, 3], [7, 1], [8, 2], [9, 7]];
+    const data5 = [[1, 1], [2, 7], [3, 5], [4, 6], [5, 5], [6, 4], [7, 2], [8, 4], [9, 8]];
+    return (
+      <Chart renderer="canvas" className={'story-chart'}>
+        <Axis id={getAxisId('bottom')} title={'index'} position={Position.Bottom} />
+        <Axis
+          id={getAxisId('left')}
+          title={KIBANA_METRICS.metrics.kibana_os_load[0].metric.title}
+          position={Position.Left}
+          tickFormat={(d) => Number(d).toFixed(2)}
+          domain={{ min: 0, max: 15 }}
+        />
+        <Axis
+          id={getAxisId('left group b')}
+          groupId={getGroupId('gb')}
+          title={KIBANA_METRICS.metrics.kibana_os_load[0].metric.title}
+          position={Position.Left}
+          tickFormat={(d) => Number(d).toFixed(2)}
+          hide={true}
+          domain={{ min: 0, max: 15 }}
+        />
+        <BarSeries
+          id={getSpecId('stacked bar 1')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor={0}
+          yAccessors={[1]}
+          stackAccessors={[0]}
+          data={data1}
+          yScaleToDataExtent={false}
+        />
+        <BarSeries
+          id={getSpecId('stacked bar 2')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor={0}
+          yAccessors={[1]}
+          stackAccessors={[0]}
+          data={data2}
+          yScaleToDataExtent={false}
+        />
+
+        <BarSeries
+          id={getSpecId('stacked bar A')}
+          groupId={getGroupId('gb')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor={0}
+          yAccessors={[1]}
+          stackAccessors={[0]}
+          data={data4}
+          yScaleToDataExtent={false}
+        />
+        <BarSeries
+          id={getSpecId('stacked bar B')}
+          groupId={getGroupId('gb')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor={0}
+          yAccessors={[1]}
+          stackAccessors={[0]}
+          data={data5}
+          yScaleToDataExtent={false}
+        />
+        <BarSeries
+          id={getSpecId('non stacked bar')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor={0}
+          yAccessors={[1]}
+          data={data3}
+          yScaleToDataExtent={false}
         />
       </Chart>
     );